requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
openai
jinja2
google-generativeai